# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type GameInputApocalypseType = typeof GameInputApocalypseType[keyof typeof GameInputApocalypseType];


export const GameInputApocalypseType = {
  ai_uprising: 'ai_uprising',
  nuclear_war: 'nuclear_war',
  virus_zombie: 'virus_zombie',
  natural_anomaly: 'natural_anomaly',
  experiment: 'experiment',
} as const;

export interface GameInput {
  apocalypseType: GameInputApocalypseType;
  /**
     * @minimum 5
     * @maximum 11
     */
  playerCount: number;
}

export interface PlayerStats {
  /**
     * @minimum 1
     * @maximum 10
     */
  survival: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  social: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  physical: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  mental: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  utility: number;
}

export interface Player {
  id: number;
  profession: string;
  gender: string;
  age: number;
  orientation: string;
  health: string;
  hobby: string;
  phobia: string;
  character: string;
  additionalInfo: string;
  knowledge: string;
  baggage: string;
  actionCard: string;
  conditionCard: string;
  motto: string;
  accentColor: string;
  secretConnection: string;
  stats: PlayerStats;
}

export interface ReshuffleInput {
  players: Player[];
}

export interface ReshuffleResult {
  players: Player[];
}

export interface Apocalypse {
  type: string;
  title: string;
  description: string;
  threatLevel: string;
  timeToImpact: string;
  survivors: string;
}

export type RoomStatus = typeof RoomStatus[keyof typeof RoomStatus];


export const RoomStatus = {
  operational: 'operational',
  damaged: 'damaged',
  critical: 'critical',
  destroyed: 'destroyed',
} as const;

export interface Room {
  name: string;
  status: RoomStatus;
  description: string;
}

export interface Resource {
  name: string;
  amount: string;
  unit: string;
}

export interface Bunker {
  name: string;
  location: string;
  capacity: number;
  duration: string;
  area: string;
  rooms: Room[];
  resources: Resource[];
  specialFeature: string;
}

export interface GameData {
  apocalypse: Apocalypse;
  bunker: Bunker;
  players: Player[];
}