# lib/api-zod/src/generated/types/apocalypse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Apocalypse {
  type: string;
  title: string;
  description: string;
  threatLevel: string;
  timeToImpact: string;
  survivors: string;
}