# lib/api-zod/src/generated/types/bunker.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Resource } from './resource';
import type { Room } from './room';

export interface Bunker {
  name: string;
  location: string;
  capacity: number;
  duration: string;
  area: string;
  rooms: Room[];
  resources: Resource[];
  specialFeature: string;
}