# lib/api-zod/src/generated/types/gameData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Apocalypse } from './apocalypse';
import type { Bunker } from './bunker';
import type { Player } from './player';

export interface GameData {
  apocalypse: Apocalypse;
  bunker: Bunker;
  players: Player[];
}