# lib/api-zod/src/generated/types/gameInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameInputApocalypseType } from './gameInputApocalypseType';

export interface GameInput {
  apocalypseType: GameInputApocalypseType;
  /**
     * @minimum 5
     * @maximum 11
     */
  playerCount: number;
}