# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerStats } from './playerStats';

export interface Player {
  id: number;
  profession: string;
  gender: string;
  age: number;
  orientation: string;
  health: string;
  hobby: string;
  phobia: string;
  character: string;
  additionalInfo: string;
  knowledge: string;
  baggage: string;
  actionCard: string;
  conditionCard: string;
  motto: string;
  accentColor: string;
  secretConnection: string;
  stats: PlayerStats;
}