# lib/api-zod/src/generated/types/playerStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerStats {
  /**
     * @minimum 1
     * @maximum 10
     */
  survival: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  social: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  physical: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  mental: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  utility: number;
}